create pages directory + move Shop
